Change option arg zoom type to be double instead of int

+++ Options.cs
@@ -15,9 +15,9 @@ namespace Mandelsharp {
         public int Y { get; set; } = -1;
 
         [Option('z', "zoom", Required = false, HelpText = "Set zoom value.")]
-        public int Zoom { get; set; } = -1;
+        public double Zoom { get; set; } = -1;
 
         [Option('c', "czoom", Required = false, HelpText = "Set cartasian zoom value.")]
-        public int CartasianZoom { get; set; } = -1;
+        public double CartasianZoom { get; set; } = -1;
     }
 }
